Skoda Mini Checkpoint
Added a pre-processor for the Skoda Mini Checkopint data set: http://www.ife.ee.ethz.ch/research/activity-recognition-datasets.html.
This time i haven't implemented a loader, since the command line used to download the .zip file doesnt work for the provided link.
Since there is only one subject in this data set I used the arms, which are separated in different data sets, for indexing. Classifying this data set is not possible with the leave one out method that is implemented in the standardized classifier.