HCI pre-processor
Added a pre-processor for the HCI data set http://www.ife.ee.ethz.ch/research/activity-recognition-datasets.html . Data set is similar to the Skoda Mini Checkpoint in a way, that it only has one subject with various measurements done with that subject, making it not usable in a leave one out scenario of the standard classifier. The data set is pre-processed the same way as the other with the exception that is indexed by the type of movement and not the subject id.